Refactored Code organization(structure)
