+++ src/Frames/GameFrame.java
@@ -8,7 +8,7 @@ import java.awt.*;
 import java.awt.event.MouseEvent;
 import java.awt.event.MouseListener;
 
-public class GameFrame extends JFrame implements MouseListener {
+public class GameFrame extends JFrame{
     public static Dimension SCREEN_SIZE = new Dimension(700, 442);
     public GameFrame(){
         this.setIconImage(new ImageIcon("assets/images/MINESWEEPER.PNG").getImage());
@@ -21,31 +21,6 @@ public class GameFrame extends JFrame implements MouseListener {
 
         this.setLocationRelativeTo(null);
         this.setVisible(true);
-
-    }
-
-    @Override
-    public void mouseClicked(MouseEvent e) {
-
-    }
-
-    @Override
-    public void mousePressed(MouseEvent e) {
-
-    }
-
-    @Override
-    public void mouseReleased(MouseEvent e) {
-
-    }
-
-    @Override
-    public void mouseEntered(MouseEvent e) {
-
-    }
-
-    @Override
-    public void mouseExited(MouseEvent e) {
 
     }
 }
